Adding latest refactor changes based on discussion with dev team today.  Noted Jira tickets that were added to track future changes.
